requirements.txt
openai==2.1.0
pyttsx3==2.99
pyaudio==0.2.14
webrtcvad==2.0.10
python-dotenv==1.1.1
setuptools==80.9.0
awscrt==0.23.5
awsiotsdk==1.23.0
